[GEF] Streamline Handle lookup in GEF viewer
The Handles of a viewer should be determined via the findHandleAt(Point)
method declared by the GraphicalViewer interface. This interface needs
to be implemented by our GraphicalViewer, just like the Handle interface
needs to be implemented by our Handle.

The drag-tracker tool getter and setter have to be renamed to match the
signature specified by the interface.