[GEF] TargetingTool#getExclusionSet() should return a IFigure's
This change adapts the return type of the getExclusionSet() method from
a collection of EditParts to a collection of IFigures, to match the
signature of the GEF tool

As a result, the findTargetEditPart() method needs to be adapted to now
accept this collection of IFigures.

Note that this causes issues with the TreeViewer, as it is based on
TreeItems, not IFigures. To solve this, the exclusion set has been moved
and is now a part of the Conditional.