Remove AMS component tests
This removes the test cases for the VarmenuConstraints and
VarmenuLayout.

At some point, support for those classes was provided via the
org.eclipse.wb.swing.asm plugin. However, this plugin was removed when
WindowBuilder was donated to the Eclipse foundation. Those tests
therefore cover functionality that is not provided by WindowBuilder