Reactivate BindingsTest and adapt Bindings classes to expected behavior
Whenever a new method is added to the IMethodBinding, IPackageBinding or
the ITypeBinding, the Designer implementation must be either implemented
properly or throw an IllegalArgumentException().

Instead, those methods have been implemented by returning a default
value, which brings the risk of diverging from the value of the
underlying binding and thus leading to inconsistent behavior.

The affected methods have been adapted to take this expected behavior
into consideration and for trivial cases, been correctly implemented.

Affected commits:
- cf88ae1c0c0837c7e33868bfdc06aebe2a787219
- 6912838e8f92e2033f9cd26a7b0269a662c234b2
- 6648a1ca4e007c08f2645d44bdcbb719e717ca64
- a6af2a248562552c6fd3394f35dd81a666842e1f
- 9e522234c54ec85d55288cea11a5c8e6766fcc4d
- 2849290a635fe4c10b318520bac7d81473343f84