Improve HighDPI support for Swing designer
The current implementation always takes a screenshot of the Swing
components at 100% zoom, which is then up-scaled to match the display
zoom. Doing so results in blurry and low-quality images.

Instead, the screenshot is taken at the current zoom level and contained
in an ImageDataProvider. When the image is painted this then allows us
to use the correctly scaled image data. Note that artifical scaling is
still done, if a different zoom level is requested.

Closes https://github.com/eclipse-windowbuilder/windowbuilder/issues/506