Stabilize SourceCompositeTests to avoid occasional SEGFAULT
This replaces the simulated key events generated by SWTBot with more
stable counterparts. Instead of a "ENTER" key, the active cell editor is
closed by calling "applyEditorValue" on the underlying table viewer and
instead of posting the remaining key events to "Display.post()", the
events are instead directly passed to the targeted widget.

Closes https://github.com/eclipse-windowbuilder/windowbuilder/issues/1096