[Draw2D] Determine figure under cursor using Draw2D API
The TargetFigureFindVisitor is only used in the EventManager in order to
determine the figure underneath the mouse event. However, this figura
can also be calculated using the findFigureAt() method.

Given that this is the last place where this visitor is used, the
visit() method of the Figure has become obsolete and will be deleted
together with the visitor class.