Migrate TextFeedback to use plain Draw2D labels
This adapts the VerticalLabel to extend the Draw2D Label class, rather
than our copy, in order to clean up the usage in the TextFeedback class.

There is no benefit to using our classes over the upstream classes and
migrating makes it possible to remove them at a later stage.