Repair and execute PropertyTableTooltipTest
The tool-tip tests for the property table are no longer disabled and are
therefore executed as part of the automatic tests.

During the reactivation, it was discovered that one specific behavior no
longer occurs; If the property text is longer than the column, a
tool-tip should be shown, containing the entire text (Similar to how an
SWT table behaves).