Improve access to JDT preference store
This replaces the reflective access to the JDT UI preference store with
a call to `PreferenceConstants.getPreferenceStore()`.

In addition, the PreferenceAdapter is marked as deprecated, to resolve
the warnings produced by referencing the deprecated Eclipse Preferences.
The class sadly can't be removed, because JDT Core doesn't have a public
accessor for its preference store and because this deprecated class is
required for the JavaTextTools.