Move re-layout of SWT container to EmbeddedSwingComposite
The Swing preview is created from the EventDispatcher thread, which is
different from the SWT thread. Depending on the timing, this might mean
that the layout done before the preview is created.

With this change, the layout is now always done after the Swing
components have been created, avoiding this problem. But unlike the
previous implementation, this layout must be scheduled asynchronously,
to avoid both UI threads being blocked at the same time.