Remove obsolete "ClassLoader" attribute for LookAndFeel UIManager
It's still not clear why this property is set or whether it's even
supported, but it seems like this is related to an ancient Java 1.2 bug:

https://bugs.openjdk.org/browse/JDK-4155617

To summarize, with Java 1.2.1, the LookAndFeel was calculated using:
- Class.forName(className)

Which makes it impossible load custom LookAndFeels that are not on the
Swing classpath. The bug proposes a solution to this problem by setting
the "ClassLoader" attribute.

However, this proposal seems to have been discarded. With Java 1.2.2,
the following snippet was used to the LookAndFeel:
- Class.forName(className, true, ClassLoader.getSystemClassLoader())

With Java 1.4.0, this the system class-loader was replaced with the
context class-loader, which is what's still used today.

Furthermore, the class-loader is only relevant when the fully qualified
class-name is given. But WindowBuilder supplies an instance of the
LookAndFeel, meaning that even if this mechanism were to exist, it is
not applied here.