Move access to "focusable" property to AWT event thread
The "focusable" property is stored as a local variable in the AWT
context. This field is updated when the AWT composite is initially
created and when a "focusable" event is fired.

Both operations are executed from within the AWT event thread, but the
access to this field may be from both the SWT and the AWT event thread.