Move calculation of embedded Swing component bounds to AWT thread
The size of a JComponent must not be done inside the SWT UI thread.
Otherwise this might cause a deadlock, depending on when e.g. the layout
is applied.

To keep both sizes in sync, a control listener needs to be added to the
composite which asynchronously updates the bounds of the AWT component.
Because those components are used for the layout of the SWT component, a
re-layout must be requested (also asynchronously) after the AWT bounds
have been calculated.