Move setting of LookAndFeel to AWT event thread
The LookAndFeel is no longer set in the SWT UI thread. More
specifically, it is set before the Swing preview is created and
afterwards restored.

All of those updates are done asynchronously, so this should greatly
reduce the change of a deadlock due to both threads waiting for one
another.