Deprecate support for Applet/JApplet API #383
This formally deprecates the support for java.applet.Applet and
javax.swing.JApplet in WindowBuilder. When such a container is used, a
"Deprecated" label will be shown in the components tree, in addition to
the deprecation warnings already generated by JDT.

All internal references (excluding the embedded previews) and imports
have been removed and where necessary replaced by their fully qualified
names.

In places where we check whether a container is an instance of such an
Applet, the class is dynamically loaded. If the class can't be found,
those check silently fail, rather than causing runtime errors.

Closes https://github.com/eclipse-windowbuilder/windowbuilder/issues/383