Encapsulate the Swing LookAndFeel using an ObservableValue
Access to the Swing objects must be done from the AWT event dispatcher
thread (aka the Swing UI thread). Rather than explicitly checking
whether any method is called from this thread via e.g. the EventQueue,
we can take advantage of the "Realm" concept provided by the Eclipse.

The LookAndFeel is stored within an ObservableValue. The realm of this
value is the Swing UI thread. Modifying this value from any other thread
is guaranteed to throw an exception.