Encapsulate the Swing SpinnerModel using an ObservableValue
The SpinnerModel shown in the SpinnerModelDialog is currently updated in
the SWT UI thread. On MacOS, this may lead to a deadlock, in case the
SWT UI thread is waiting for the AWT event dispatch thread.

Access to this model is now guarded by an ObservableValue, which
requires proper synchronization. The dialog has been restructured to
make sure modifications to this value are done exclusively in the event
dispatch thread and that the dialog is only updated in the SWT UI
thread.

May close
https://github.com/eclipse-windowbuilder/windowbuilder/issues/1204