Properly reset the EmbeddedSwingComposite when disposed
This restores the UIDefaults for the fonts when the widget is disposed.
Otherwise changes to e.g. the LookAndFeel fonts would be preserved, even
when the LookAndFeel dialog has been closed.

Furthermore, disposing the Swing Frame is now done via a DisposeEvent.
The dispose() method is not called recursively, so there is no guarantee
that it is actually disposed.