Properly initialize fonts in FontPropertyEditorTest
The test currently depends on the default font that is set by Swing.
This font might be changed at any point by a different test, leading to
failures that are difficult to track down.

To make the test more stable, explicitly set the expected font at the
start of the test.