Encapsulate Swing Border using a BorderValue class
This stores the Swing Border in an ObservableValue. This value is
created over the Swing Realm, meaning that all access to this instance
must be done from the AWT event dispatcher thread. The instance is
read-only and may be created from any thread.

Note that there are a few instances where the Border value is accessed
without synchronization. This will be resolved in an upcoming commit and
is done as an intermediate solution for the sake of simplicity.