Correct behavior when a Border is managed by multiple BorderComposites
This is a follow-up to the lookup table that was introduced
with870f5bdfbf1c90e31288fe1723e839532497a5f8. In some cases it might
happen that the SwingBorderComposite is used to manage classes that
should be managed by one of the other composites (e.g. EmptyBorder).

Because a simple HashMap is used, we can't guarantee that the
SwingBorderComposite is checked last. Because of this, it must not be
considered when iterating over the elements in this map and instead be
explicitly checked afterwards.

Additionally, the check in the EmptyBorderComposite has been extended to
allow subclasses of the EmptyBorder, as long as it's not a subclass of
the MatteBorder, which is managed by the MatteBorderComposite.