Fix NPE when restoring layout to default value
The changes introduced with 7ff0657587397a86147015cd37cd19dd51afac6e
cause an exception, if no layout is set. Instead of an empty string, the
"null" string is used to describe the implict layout. This breaks the
selection on the preference page.

If no layout with this id is found, the call to setDefaultLayout() fails
with a NullPointerException, which is the case for the default selection
or if a layout no longer exists.

Using the GridLayout as an implicit layout is also not the correct
behavior. If this is the desired behavior, the GridLayout must be
selected in the preference page.