Partially revert 87d34be1a1fafe9ffbd59289b4e8017247f321ec
As part of this commit, the menu decorations are no longer drawn on
Linux when the SHELL_TRIM style is set. The original motivation was to
better handle the offsets when a menu is set.

However, this now causes offset errors because all widgets are
positioned relative to the top left corner of the shell, but the bounds
of the shell still takes the decorations into consideration.