Simplify SWT screenshot logic on GTK/Linux
The current implementation of taking a screenshot for each widget starts
with the window for the shell and then traverses along child windows. In
case a widget is registered to multiple windows, the same screenshot is
taken multiple times, leading to inefficiencies.

Main motivation for this approach seems to be that originally, the
screenshot logic was fully implemented in C, which doesn't have access
to the high-level widget structure.

Instead of iterating over the windows, we now start with the shell
widget and then recursively traverse through all of its children.