DesignerEditorTestCase should extend AbstractJavaInfoTest
Some tests are occasionally failing because the EditorState is still
initialized with a JavaInfo object that was created during a previous
test. By extending the AbstractJavaInfoTest class, this class is now
explicitly cleared after each test.

Semantically, this change also sound; The AbstractJavaInfoRelatedTest is
used for test cases that don't parse JavaInfo objects. However, almost
all subclasses of the DesignerEditorTestCase do so.