Make dependency from test plugin to OS support implementation explicit
The latest version of PDE calculates the minimal set of required plugins
when executing a test case. Because there is no clear dependency from
"org.eclipse.wb.so" to its platform-specific implementations, only an
incomplete set of plugins is selected, which leads to an OSSupportError
during execution.

We can add them directly due to their platform filter. So instead the
dependency is described via the "Provide-Capability" and
"Require-Capability" OSGi headers.