Disable BurningWave initialization by default #1305
BurningWave is using Unsafe to break the module system introduced with
Java 9, in order to allow WindowBuilder to access the internal Swing
fields and methods.

The implementation has already been adapted in the past so that this
dependency can be disabled, by setting the "wbp.burningwave.enabled"
system property to "false". But for the sake of backwards compatibility,
this property was initialized with "true".

With this change, this property now needs to be explicitly enabled.

Closes https://github.com/eclipse-windowbuilder/windowbuilder/issues/1305