Process deferred layout requests before taking screenshot #1324
When the "expanded" property is set for an ExpandableComposite, the
"isLayoutDeferred()" property is set as well.

This causes the layout to be done _after_ the screenshot has been taken,
meaning that at that point it still has an empty size.
As a result, there is no image to show in the designer page.

Closes https://github.com/eclipse-windowbuilder/windowbuilder/issues/1324